Port D1 Activities to React
