--- src/app/activities/activities.html
@@ -1,58 +1,0 @@
-<div class="activities dim-page">
-
-  <!-- TODO: need a more generic character-tile directive -->
-  <div class="activities-characters">
-    <div class="activities-character" ng-repeat="store in $ctrl.stores | sortStores:$ctrl.settings.characterOrder track by store.id">
-      <div class="character" ng-class="{ current: store.current }">
-        <div class="character-box">
-          <div class="background" ng-style="{ 'background-image': 'url(' + store.background + ')' }"></div>
-          <div class="details">
-            <div class="emblem" ng-style="{ 'background-image': 'url(' + store.icon + ')' }"></div>
-            <div class="character-text">
-              <div class="top">
-                <div class="class">{{ ::store.className }}</div>
-                <div class="powerLevel" ng-if="::!store.isVault">{{ store.powerLevel }}</div>
-              </div>
-              <div class="bottom">
-                <div class="race-gender" ng-if="::!store.isVault">{{ ::store.genderRace }}</div>
-                <div class="level" ng-if="::!store.isVault">{{ store.level }}</div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-
-  <div class="activity section" ng-repeat="activity in $ctrl.activities track by $index">
-    <div class="title activity-header" ng-style="::activity.image | bungieBackground" ng-class="{ 'activity-featured': activity.featured }"
-      ng-click="$ctrl.toggleSection('activities-' + activity.hash)">
-      <span class="collapse-handle">
-        <font-awesome-icon class="collapse" icon="expand" ng-if="$ctrl.settings.collapsedSections['activities-' + activity.hash]"></font-awesome-icon>
-        <font-awesome-icon class="collapse" icon="collapse" ng-if="!$ctrl.settings.collapsedSections['activities-' + activity.hash]"></font-awesome-icon>
-
-        <img class="small-icon" ng-src="{{ ::activity.icon | bungieIcon }}" />
-        <span class="activity-name">{{ ::activity.name }}</span>
-
-        <font-awesome-icon icon="star" ng-if="::activity.featured"></font-awesome-icon>
-      </span>
-      <span class="activity-type" ng-bind="::activity.type"></span>
-    </div>
-
-    <div class="activity-info" ng-if="!$ctrl.settings.collapsedSections['activities-' + activity.hash]">
-      <div class="activity-progress" ng-repeat="tier in activity.tiers track by $index">
-        <div class="tier-title" ng-if="activity.tiers.length > 1">{{ ::tier.name }}</div>
-        <div class="tier-characters">
-          <div class="tier-row" ng-if="tier.characters.length" ng-repeat="character in tier.characters | sortStores:$ctrl.settings.characterOrder track by character.id">
-            <span class="step-icon" ng-class="{ complete: step.complete }" ng-repeat="step in character.steps track by $index"></span>
-          </div>
-        </div>
-      </div>
-
-      <div class="activity-skulls" ng-if="::activity.skulls" ng-repeat="skull in activity.skulls">
-        <img class="small-icon" ng-src="{{ ::skull.icon | bungieIcon }}" />
-        {{ ::skull.displayName }}<span class="weak"> - {{ ::skull.description }}</span>
-      </div>
-    </div>
-  </div>
-</div>
