--- src/app/activities/activities.module.ts
@@ -1,5 +1,0 @@
-import { module } from 'angular';
-
-import { ActivitiesComponent } from './activities.component';
-
-export default module('activitiesModule', []).component('dimActivities', ActivitiesComponent).name;
