+++ src/app/activities/routes.ts
@@ -1,16 +1,10 @@
 import { ReactStateDeclaration } from '@uirouter/react';
-import { angular2react } from 'angular2react';
-import { lazyInjector } from '../../lazyInjector';
-import { ActivitiesComponent } from './activities.component';
+import Activities from './Activities';
 
 export const states: ReactStateDeclaration[] = [
   {
     name: 'destiny1.activities',
     url: '/activities',
-    component: angular2react(
-      'dimActivities',
-      ActivitiesComponent,
-      lazyInjector.$injector as angular.auto.IInjectorService
-    )
+    component: Activities
   }
 ];
