+++ src/app/destiny1/destiny1.module.ts
@@ -1,11 +1,10 @@
 import { module } from 'angular';
 
-import activitiesModule from '../activities/activities.module';
 import loadoutBuilderModule from '../loadout-builder/loadout-builder.module';
 import vendorsModule from '../vendors/vendors.module';
 
 export { states } from './routes';
 
-const mod = module('destiny1Module', [activitiesModule, loadoutBuilderModule, vendorsModule]);
+const mod = module('destiny1Module', [loadoutBuilderModule, vendorsModule]);
 
 export const angularModule = mod;
